+++ src/main/commands/exists_async.cc
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright 2013-2018 Aerospike, Inc.
+ * Copyright 2013-2020 Aerospike, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -64,4 +64,5 @@ NAN_METHOD(AerospikeClient::ExistsAsync)
 Cleanup:
 	delete cmd;
 	if (key_initalized) as_key_destroy(&key);
+	if (p_policy && policy.base.predexp) as_predexp_list_destroy(policy.base.predexp);
 }